Extracted the filter
