+++ code/app/src/main/java/com/example/y/views/FollowingMoodEventListActivity.java
@@ -16,7 +16,7 @@ public class FollowingMoodEventListActivity extends MoodListActivity {
 
         // Initialize controller
         controller = new FollowingMoodListController(this, unused -> {
-            moodListView.setAdapter(controller.getMoodAdapter());
+            onControllerInitialized();
         }, this::handleException);
     }
